refactor entry
